+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/bootstrap/AzureMonitorMeterRegistryFactory.java
@@ -1,23 +1,19 @@
-package uk.gov.justice.services.metrics.micrometer.azure.producers;
+package uk.gov.justice.services.metrics.micrometer.bootstrap;
 
 import static io.micrometer.core.instrument.Clock.SYSTEM;
 
 import uk.gov.justice.services.metrics.micrometer.azure.config.FrameworkAzureMonitorConfig;
 
-import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
-import javax.inject.Singleton;
 
 import io.micrometer.azuremonitor.AzureMonitorMeterRegistry;
 
-@Singleton
-public class AzureMonitorMeterRegistryProducer {
+public class AzureMonitorMeterRegistryFactory {
 
     @Inject
     private FrameworkAzureMonitorConfig frameworkAzureMonitorConfig;
 
-    @Produces
-    public AzureMonitorMeterRegistry meterRegistry() {
+    public AzureMonitorMeterRegistry create() {
         return new AzureMonitorMeterRegistry(
                 frameworkAzureMonitorConfig,
                 SYSTEM);