+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/config/MetricsConfiguration.java
@@ -1,7 +1,6 @@
 package uk.gov.justice.services.metrics.micrometer.config;
 
 import static java.lang.Boolean.parseBoolean;
-import static java.lang.String.format;
 
 import uk.gov.justice.services.common.configuration.Value;
 
@@ -9,15 +8,12 @@ import javax.inject.Inject;
 
 public class MetricsConfiguration {
 
-    private static final String AZURE_METRICS_MONITOR_CONNECTION_STRING = "azure.metrics.monitor.connection.string";
-    private static final String MICROMETER_METRICS_ENABLED = "micrometer.metrics.enabled";
-
     @Inject
-    @Value(key = MICROMETER_METRICS_ENABLED, defaultValue = "false")
+    @Value(key = "micrometer.metrics.enabled", defaultValue = "false")
     private String micrometerMetricsEnabled;
 
     @Inject
-    @Value(key = AZURE_METRICS_MONITOR_CONNECTION_STRING, defaultValue = "azure-metrics-connection-string-not-set")
+    @Value(key = "azure.metrics.monitor.connection.string", defaultValue = "azure-metrics-connection-string-not-set")
     private String azureMonitorConnectionString;
 
     public boolean micrometerMetricsEnabled() {
@@ -25,11 +21,6 @@ public class MetricsConfiguration {
     }
 
     public String azureMonitorConnectionString() {
-
-        if(azureMonitorConnectionString == null) {
-            throw new MicometerMetricsConfigurationException(format("Azure monitor connection string found in JNDI. Value should be set for the name '%s'", AZURE_METRICS_MONITOR_CONNECTION_STRING));
-        }
-
         return azureMonitorConnectionString;
     }
 }