+++ metrics/metrics-micrometer-core/src/main/java/uk/gov/justice/services/metrics/micrometer/counters/EventsFailedCounterMeter.java
@@ -1,7 +1,6 @@
 package uk.gov.justice.services.metrics.micrometer.counters;
 
 import static uk.gov.justice.services.metrics.micrometer.meters.MetricsMeterNames.EVENTS_FAILED_COUNTER_NAME;
-import static uk.gov.justice.services.metrics.micrometer.meters.MetricsMeterNames.EVENTS_IGNORED_COUNTER_NAME;
 
 import uk.gov.justice.services.metrics.micrometer.meters.MetricsMeter;